Integrated Search Functionality

+++ main.py
@@ -41,23 +41,39 @@ def save():
         field_empty = messagebox.showinfo(title="Oops", message="Some of the fields are empty")
 
     else:
-        #save_info = messagebox.askyesno(title=website, message=f"These are details: \nemail: {email}\npassword: {password}\nAre you sure to proceed?")
-        #if save_info:
+        save_info = messagebox.askyesno(title=website, message=f"These are details: \nemail: {email}\npassword: {password}\nAre you sure to proceed?")
+        if save_info:
 
-        try:
-            with open("Password_Manager_Data.json", "r") as data_file:
-                data = json.load(data_file)
-                data.update(new_data)
-            with open("Password_Manager_Data.json", "w") as data_file:
-                json.dump(data, data_file, indent=4)
+            try:
+                with open("Password_Manager_Data.json", "r") as data_file:
+                    data = json.load(data_file)
+                    data.update(new_data)
+                with open("Password_Manager_Data.json", "w") as data_file:
+                    json.dump(data, data_file, indent=4)
 
-        except FileNotFoundError:
-            with open("Password_Manager_Data.json", "w") as data_file:
-                json.dump(new_data, data_file, indent=4)
+            except FileNotFoundError:
+                with open("Password_Manager_Data.json", "w") as data_file:
+                    json.dump(new_data, data_file, indent=4)
 
-        finally:
-            website_entry.delete(first=0, last=END)
-            password_entry.delete(first=0, last=END)
+            finally:
+                website_entry.delete(first=0, last=END)
+                password_entry.delete(first=0, last=END)
+
+
+# ---------------------------- SEARCH DATA ------------------------------- #
+
+def search_data():
+    with open("Password_Manager_Data.json", "r") as data_file:
+        data = json.load(data_file)
+
+    website = website_entry.get()
+    if website in data:
+        email = data[website]["email"]
+        password = data[website]["password"]
+
+        email_entry.delete(first=0, last=END)
+        email_entry.insert(0, email)
+        password_entry.insert(0, password)
 
 # ---------------------------- UI SETUP ------------------------------- #
 
@@ -74,9 +90,12 @@ canvas.grid(column=1, row=0)
 website_label = Label(text="Website")
 website_label.grid(column=0, row=1, sticky="w")
 
-website_entry = Entry(width=35)
+website_entry = Entry(width=21)
 website_entry.focus()
-website_entry.grid(column=1, row=1, columnspan=2, pady=5, sticky="w")
+website_entry.grid(column=1, row=1, sticky="w")
+
+search_button = Button(text="Search", width=15, command=search_data)
+search_button.grid(column=1, row=1, columnspan=2, pady=5, sticky="e")
 
 email_label = Label(text="Email/Username")
 email_label.grid(column=0, row=2, sticky="w")
@@ -92,7 +111,7 @@ password_entry = Entry(width=21)
 password_entry.grid(column=1, row=3, sticky="w")
 
 generate_button = Button(text="Generate Password", command=generate_password)
-generate_button.grid(column=1, row=3, columnspan=2,pady=5, sticky="e")
+generate_button.grid(column=1, row=3, columnspan=2, pady=5, sticky="e")
 
 add_button = Button(text="Add", command=save, width=36)
 add_button.grid(column=1, row=4, columnspan=2, pady=5)
